requirements.txt
telethon==1.34.0
python-telegram-bot==20.7
asyncio==3.4.3
aiofiles==23.2.1
APScheduler==3.10.4
cryptography==41.0.8
PyYAML==6.0.1
psutil==5.9.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
watchdog==3.0.0
colorama==0.4.6
requests==2.31.0